Fixed single word !chat about command
This fix was pretty simple once I noticed what to look for.
The earlier code used each_key(word), however each_key doesn't take any
arguments in ruby 1.9.3. This causes some strange problems which I only
saw in the rbot.log file.

Now since I am not able to filter the keys on my keyword I had to remove
the break block as well. Since I have to iterate over everything. This
will slow down the bot in time, but the filtering does take its toll to be
fair.

I will look for a better solution, but this works for the moment.